% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-relations.R
\name{bench_manifest_paths}
\alias{bench_manifest_paths}
\title{List benchmark relation manifests}
\usage{
bench_manifest_paths(
  path = system.file("extdata", package = "VariantStoryBench")
)
}
\arguments{
\item{path}{Directory containing CSV relation manifests.}
}
\value{
Sorted CSV paths.
}
\description{
List benchmark relation manifests
}
