% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-classes.R
\name{BenchCommandAdapter}
\alias{BenchCommandAdapter}
\title{A command-line benchmark adapter}
\usage{
BenchCommandAdapter(
  executable = character(0),
  arguments = character(0),
  working_directory = NULL
)
}
\arguments{
\item{executable}{Command executable.}

\item{arguments}{Command arguments. \code{{input}} and \code{{output}} are replaced
with the paths passed to \code{\link[=bench_execute]{bench_execute()}}.}

\item{working_directory}{Optional command working directory.}
}
\description{
A command-line benchmark adapter
}
