% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-results.R
\name{bench_manifest_paths}
\alias{bench_manifest_paths}
\title{List benchmark result manifests}
\usage{
bench_manifest_paths(
  path = system.file("extdata", package = "VariantStoryBench")
)
}
\arguments{
\item{path}{Directory containing CSV result manifests.}
}
\value{
Sorted CSV paths.
}
\description{
List benchmark result manifests
}
