% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{VariantStoryBench-package}
\alias{VariantStoryBench}
\alias{VariantStoryBench-package}
\title{VariantStoryBench: reproducible variant reanalysis benchmarks}
\description{
Tools for defining benchmark results, executing declared command adapters,
and calculating ranking metrics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RGenomicsETL/VariantStoryBench}
  \item Report bugs at \url{https://github.com/RGenomicsETL/VariantStoryBench/issues}
}

}
\author{
\strong{Maintainer}: Sounkou Mahamane Toure \email{sounkoutoure@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Sounkou Mahamane Toure \email{sounkoutoure@gmail.com} [copyright holder]
}

}
\keyword{internal}
